Do a no-microservices conversion for RGWCopyObj.
We shouldn't ever see RGWCopyObj / copy-object when running with the
microservices platform.

However, when testing our authz changes, it's useful to be able to use
copy-object. Here we've used some new features to fail hard normally (assuming
the µservices) and to allow the copy-object operation when we've explicitly
set up RGW to run in the non-µservices mode.

Note that required nerfing RGWPutACLs in handoff_authz mode - s3cmd will do a
put-object-acl after a copy-object, and we need to 'allow' that even if we
don't set any actual ACLs. It just returns success and does nothing if Handoff
Authz is enabled.