Note a suspected (serious!) bug in delete-object.
As the comment says, I think the check is wrong - it appears to check whether
the request has a version ID in it, so if you send a request that doesn't have
a version ID (e.g. s3cmd rm) it will not even check the retention or legal
hold, it'll just delete the object.

Very serious, if correct.