Allow (by default) copy-object in gen2.
It was explained to me that copy-object is often used to change tags despite
the fact that objects in S3 are immutable. copy-object acts like a UNIX
filesystem hard link, and creates a new head pointing to a chain of RADOS
objects. It's really quick.

This is apparently a frequent enough occurrence that we're going to allow
copy-object through envoy unmolested, so we can no longer block it and return
INVALID REQUEST. Wrap this decision in a config bool (defaulting to 'handle
copy object normally') so we can test easily and if necessary change
direction.