Catch all C++ exceptions in Stategy::authenticate().
We don't know why yet, but it seems an exception in get_auth_data()
isn't being caught. Speculate that it's a toolchain problem and catch
all exceptions with a log message and return EACCES.

If this doesn't catch it, we have more debugging to do. If it does, we
have to identify the exact exception (probably with more logging or
maybe a debug attachment) and then figure out why if it's an exception
of type const int that we're not catching it.