Wrap __cxa_throw in librgw_common and add test.
Add a wrapper function for __cxa_throw (libstdc++'s exception handler)
and wrap it in librgw_common. It appears we don't need to wrap it in the
target binaries, just in the library. At least this is true on my Fedora
dev workstation, we'll see on CentOS Stream.

Add a unittest that throws normal exceptions, nested exceptions, and
mimics Vitaly's standalone test binary. The first two are run many times
and check that we have actually wrapped the exception handler (in a very
crude way, by checking that a counter has incremented - that's why we
run them many times.)

The last test is reworked as a Gtest 'death test', where gtest asserts
that the program has actually died as a result. This is useful in itself
as we make sure we've modeled something that will actually crash the
program, and it's also useful to give us a way to test the debugger
attachment into our __cxa_throw wrapper which we'll need to debug rgw
proper.