Unbreak handoff ChannelRecoversFromDeadAtStartup test.
We recently fixed HandoffGRPCChannel::get_default_channel_args() which
was returning an empty ChannelArguments by mistake. The fix broke unit
test
  TEST_F(HandoffHelperImplGRPCTest, ChannelRecoversFromDeadAtStartup)

because of the way grpc::ChannelArguments is implemented. Despite
feeling like it should be a map<string, variant> it's actually a list of
keys with a variant-type value. Setting a key just adds the key to the
end of the list, which is sure to be quicker than implementing a map,
but it means that there's no uniq'ing of the key. Worse, gRPC doesn't
guarantee which of these repeated keys is used. For channels, it appears
to use the first one.

So when we set a channel argument, we really only want to set it once in
the lifetime of the grpc::ChannelArguments object. This works fine in
the context of the config observer that listens for our channel timer
settings, because the get_default_channel_args() method fetches the
config and sets up those variables.

It didn't work for the test, though - we used get_default_channel_args()
which set the keys for the times, then we set the keys again in the
test. (It worked before because get_default_channel_args() was bugged.)
That means we end up using the configured defaults, which are sensible
for production but far too long for the test. The test wants to set the
minimal time, so it sets 1ms (the only unit is ms), but we know from
reading the code that gRPC sets a minimum of 100ms regardless. So the
test waits for 105ms.

It's a useful test, because it caught this regression and because it's
useful to have a way to test that gRPC's reconnection mechanics work as
we expect.