Only strip HTTP_ from headers that start with it.
Not all headers are prefixed with HTTP_. When it was just x-amz- headers
received from the client directly, we could safely assume header
x-amz-foo was stored by the code as 'HTTP_X_AMZ_FOO'. Other headers are
sometimes filled in by the RGW code, such as 'Content-Length'; and some
headers are synthetic and never sent by the client such as 'REMOTE_ADDR'
and 'REQUEST_METHOD'.

In any case, we need to only strip the HTTP_ prefix if it is actually
present.