Add start_trace_with_req_state_parent().
Based closely on example code from OpenTelemetry, construct a trace context
from the traceparent and tracestate HTTP headers. This will inherit the trace
ID from its caller and set up a new span that sits within the parent span in
traceparent.

The tracestate header provides context for the trace, it's not expected to be
useful to RGW directly. We just pass it through to the OpenTelemetry API to
construct its trace object.

The example code is in src/jaegertracing/opentelemetry-cpp/examples/http (in
the Ceph source). It appears quite version-specific, and so a new Ceph version
might well break it. Hopefully the examples will be updated to match. We're
matching closely what the server.cc example file does, then adding a few bits
and pieces.

In particular, note the HttpTextMapCarrier class. This is a wrapper around a
map<string, valuetype> that the API uses to fetch string key/value pairs. We
copied the boilerplate from the example and removed the special-casing of the
HTTP header case which doesn't apply to us. When the trace is created, the
propagation API will call Get() on the carrier to fetch the traceparent and
tracestate headers. It feels a bit clumsy but it works quite well.

Once we've set up the header carrier, it's just a question of calling the
right otel APIs to set up a parent context using propagation, then setting up
a new trace with that parent context passed as an option.