Compile OpenTelemetry with OTLP gRPC support.
This is complicated. We need to send OTLP traces via the gRPC transport, as
that's all that the trace stack gen2 is using supports.

Firstly, we can rely on gRPC support now, at least in Akamai builds. So we can
do this without changing much in the overall build system. However, we have to
teach the in-tree OpenTelemetry source to do it, with the CMake
ExternalProject mechanism in the way.

This means solving a few problems:
- Patch the in-tree submodule.
- Send our CMAKE_PREFIX_PATH into the ExternalProject build. I thought we were
  doing this before, but we weren't.
- Configure OpenTelemetry to build the pieces we need.
- Manually create targets that Ceph CMake can use to compile and link against.

Patching the source is easy enough. That's the PATCH_COMMAND addition to
ExternalProject_Add. Getting the diff inside a submodule is a pain, but once
established that `git diff --submodule=diff` works, it's ok. The patch command
is straightforward enough. The diff file is in cmake/modules because that
makes sense to me (it's where BuildOpentelemetry.cmake is), and it has the
'.cmakepatch' suffix because (a) it sort of makes sense, and (b) files ending
in .diff and .patch are .gitignore'd at the top level.

The patch itself just changes the source to look for 'protobuf' instead of
'Protobuf', and makes the gRPC include REQUIRED.

The CMAKE_PREFIX_PATH thing is explained in BuildOpentelemetry.cmake. It's a
limitation of the CMake language that we can't easily pass strings that
contain semicolons - see the linked documents in the source. The solution
seems to work pretty reliably, and is recommended by Brad King who really
ought to know.

The only build option that's significant is WITH_OTLP=ON. The others are
tweaks - remove CMAKE_PREFIX_PATH for the reasons discussed; set
CMAKE_BUILD_TYPE to match the main source; hardcode the protoc executable.

The targets are so that CMake can first build the opentelemetry libraries we
need, and then so that Ceph CMake can refer to the built targets later. This
was discovered by trial and error. Note that three additional build targets
were added (opentelemetry_exporter_otlp_grpc, opentelemetry_otlp_recordable,
and opentelemetry_proto) as these are libraries needed to link. The latter is
a build of the compiled protobuf generated source files.

We also have to add some additional include directories. Most point to new
parts of the source that we're building now for OTLP, but the last one
(${RGW_GRPC_ROOT_DIR}/include/) is subtle. The protobuf and gRPC generates
sources expect standard gRPC header files to be present, and in our case these
are in /usr/local/grpc/include.