Support parent-based sampling; add configuration.
In OTEL the sampler governs when a trace is exported to an external system. We
generate traces in memory for all requests, but in production we only want
export a subset. This is well-described here:
  https://opentelemetry.io/docs/concepts/sampling/

Add configuration options:
- otlp_sampler_parent_based
- otlp_sampler_delegate_defaults_to_on

The first (on by default) sets 'parent based' sampler mode, so we accept
configuration from the parent span that governs whether or not a trace is
exported. This is to protect the tracing infrastructure from overload.

The second configures sampler behaviour if there is no parent trace, or if the
parent trace doesn't have the appropriate configuration. If true (the
default), we will sample in the absence of configuration. If false, we will
not sample.

Also: Use namespaces to try to make the tracer init a little easier to read.