Attempt to robustly patch OTel source.
Sigh. We can't use `git apply` because RPM build (and debuild for that matter)
don't preserve the .git directories - they untar a source tarball. However,
patch(1) has no robust way of handling an already-patched file. So we have to
try anyway, try to interpret the error message, and do our best. If we see
what looks like an already-applied patch we return success (though noting what
we did in the log). If we get an error and don't see 'Reversed (or previously'
then we fail.

This can be broken if a patch partially applies, or the patch file is changed.
However, that risk can be minimised by keeping patches very short (ideally one
change). It's a great deal better than failing builds all the time.