StoreQuery: objectlist and mpupload boilerplate.
Add parsing and object skeletons for new StoreQuery commands objectlist and
mpupload. These commands are used to list all objects in a bucket and to list
all multipart uploads in a bucket, respectively.

There's no attempt to return a merged list of both the objects and the
in-flight multipart uploads. The more one things about this, the less sense it
makes to try - we have a list that might take a very long time to complete,
and then we want somehow to merge in a list that is very likely to take much
less time, requiring us to keep a bunch of state. Better to make the split
explicit IMO.