Checkpoint: Working mpuploadlist with pagination.
Don't attempt to have a common list item struct. It's a fair amount of
abstraction pain for no obvious benefit. Just use a nested class for each item
type.

mpuploadlist is implemented similarly, but not identically, to objectlist.
Note the difference (as observed with objectstatus) in the way the marker is
handled. Note also the clearance of the struct each time.