Tidyup. Un-inline a few items. Indirection for MP.
More doxygen.

Un-inline the dump() calls, there's no need for them to be in the header. Rely
on the optimiser to inline them if it thinks it matters.

Add the same _impl indirection for mpuploadlist, so in principle a unit test
can be written without having to abstract a million things away. It might not
work; it might still be necessary to abstract a million things away just to
implement a subclass of the individual list items, but at least we're trying.