Pre-validate base64 parameters. Tweak error msgs.
We can quickly establish in the parser that a parameter contains valid base64.
There's no reason to go to all the trouble of setting up the operation if we
know it's going to fail.

Add some tests for this.

Change some error messages to have consistent case (lead with lower case
unless it's an abbreviation).