Checkpoint: Add base64 key encoding.
Keys can have characters that might break the JSON or cause other problems.
Keys can be any valid utf-8 string up to 1024 bytes (see
https://docs.aws.amazon.com/AmazonS3/latest/userguide/object-keys.html) but
there are recommendation that Amazon make. We should play it safe.

When speaking to Tunc, his preference was to always encode. I can see the
sense there, it makes the logic simpler at the cost of running a little
slower. I am going to remove the conditional encoding with the next commit,
but let's have a record of it in the repository.