Some additional stats. Minor cleanup.
Add sal_not_current, at Ety's request. Apparently the client will sum
all the results to make some calculation about the ratio of current
(an object is present or a delete marker) vs not-current (old versions)
that will be useful to them.

Add a count of SAL queries. This may be useful in the future.

Minor cosmetic tweaks and commentary.