Unbreak usage logging.
Usage logging was broken with Handoff Authz enabled. The problem was
that s->bucket_owner was not set in the authz path. That field is used
in src/rgw/rgw_log.cc:log_usage(), to set the logged user in the case
where the request has an associated bucket.

(That it's used in the 'associated bucket' case explains why requests
like create-bucket were still in the usage log - technically that
request doesn't have a bucket associated with it, at least not at the
start of the request. In those cases, the user ID is set from the
request.)

The fix is to set s->bucket_owner in the Handoff Authz path. In the
non-Handoff path, the owner is deduced by loading the bucket ACL.
Happily we don't have to do that, as in Handoff Authn we *always* get
the bucket owner as the user ID of the request. So we can use the
request's already-set user ID.