Check we're dealing with a handoff user.
As pointed out by vitaly@, we need to be aware of subusers, which can't be
bucket owners.

Now, in gen2, we only use authn handoff, and the Authenticator only returns
the bucket owner. This is reliable for S3 commands.

However, we might be using other users for, say, the admin API. So we'll
compromise: If the user returned is coming from handoff authn, let's assume
it's the bucket owner and proceed.

If the user is not coming from handoff authn, we don't know for sure that it's
a subuser, so execute the old (non-handoff authz) codepath, and load the
policy to discover the bucket owner.