More complete solution to handoff bucket_owner set.
Whilst I'm fairly convinced that previous approaches would have been ok,
let's play it super safe: Let's follow the logic of the non-Handoff
Authz code to set the bucket owner.

So: Trace the calls from

  read_bucket_policy()
    -> rgw_op_get_bucket_policy_from_attr()
      -> RGWAccessControlPolicy::get_default()

and do what they do. This simple version doesn't do the same error
checks, we'll add these in presently.

The end result is that we get the owner directly from the bucket, and if
that's wrong we have bigger problems.