More complete bucket owner fetch implementation.
Duplicate more of the logic discussed in commit 0258a46. Perform the
same validity checks, minus the 'empty bucket name' check which is
performed earlier in the function.

Add some level 20 logging to show the result.

Omit the TYPE_HANDOFF check. It seems that somehow the type isn't set,
which I don't understand - I *know* it's set when the applier is
created.