Add instance ID to the continuation token.
We were having problems with versioned buckets and continuation markers.
To cut a long story short, when one sets the 'marker' when issuing the
query to the cluster, it is possible to encode not only the object key
name, but also the object instance ID.

This doesn't matter for non-versioned buckets. Nor does it matter for
queries that ignore object versions, which is always the case for RGW's
own queries in RGWListObj*.

It matters to us though: We _do_ query versions (we need to see delete
markers) and so the object key name isn't enough. We need the instance
ID to start subsequent queries in the right place. Otherwise we restart
at the 'top' of the list of objects under the object key name, which is
_always_ the 'current' object. So we loop.

Solution: Re-encode the continuation token as a JSON object, and in
addition to the object key name, add the instance ID. When issuing the
query, feed both into the params object of the query. The loop no longer
occurs because we restart at the proper instance ID, not just the name.

Next: Un-static the encode and decode methods and add some unit tests.