Add some paranoia checks around the cont. token.
No prod code changes, only tests.

Check that the continuation token JSON doesn't need the keys to be base64
encoded. Rather than just add it, the tests use the production
create_continuation_token() and read_continuation_token() functions to verify
sensible behaviour when the keys have bad characters.

I'm more inclined to not worry so much because we control generation, line
coding (the base64 encoding of the whole token) and reading of the
continuation token. If someone's going to send a malicious token that's a
separate problem - we're then relying on the decoder to not crash, but that's
not something that base64-encoding could prevent (the attacker would just
base64 encode the payload). All we want to try to ensure is that if there's a
key with non-JSON characters, it will be properly encoded and decoded, and
that appears to be true.

In retrospect if that's true we may have been able to skip this for the main
JSON (the objectlist and mpuploadlist lists) as well, but we've already
crossed that bridge.