Wire up to log_usage().
Fetch the usage bypass flags once per call to log_usage(), and based on the
returned flags either log or don't log the network transfer bytes.

Added an API that fetches the flags instead of returning a bool, as there's
really no reason to parse the header twice.

Changes the options to use hyphens instead of underscores, it makes more sense
in a text field.