Basic files and CMake for mdoffload driver+unit.
The minimal CMake and C++ for the new driver, including the CMake flag
WITH_RADOSGW_MDOFFLOAD to enable (default to ON, for the sake of my
sanity).

We'll live in src/rgw/driver/, because that's where things end up living
in later versions of Ceph.

Some of this is cribbed from the dbstore driver and tweaked. Currently
none of it does anything, it's the bare minimum to get it compiled and
linked.