Add intercepts for bucket attrs. Mock SAL classes.
Lots going on here.

All the points I've found so far for bucket attrs to be set are
intercepted here. Reads are get_attr(). Writes are via
merge_and_store_attrs(). The only subtlety is that set_attr() doesn't
appear to be used - the initial set is performed in the driver's
get_bucket(dpp, user*, rgw_bucket&, rgw::sal::bucket*) method. So we
need to override <driver>::get_bucket() too to catch that.

Also: Use Copilot to generate gmock classes for many of the classes in
rgw_sal.h. (Driver, User, Bucket, Object, MultipartPart,
MultipartUpload). Took some time (mostly context load), and took a few
iterations to fix the errors, but the result looks sound.

Add basic create-and-null-check unit tests for the filter driver. Next
up is to do some mocking and try to get and set some attributes.