Add first useful tests via the mock.
Some small tweaks to the SAL mocks to make them usable. I'm sure
there'll be corresponding changes as we move to MockObject. Currently
MockDriver and MockBucket are in flight.

Move the mock classes to namespace ::akamai::mock, there's no need to
pollute ::rgw::sal.

Note that the first mock tests fail! This is ok - the point of the tests
is that they check that get_attr() and merge_and_store_attrs() don't
call the underlying implementation, which is kinda the point of the
filter driver. It's a start.

Next up is to do the same for MockObject/MDOffloadFilterObject. This
time, we can do so in a more TDD because the framework is somewhat
there.

After that, it's time to set up the infrastructure to call out for
Bucket and Object attributes.