Successfully mock <User>::create_bucket().
That was tougher than it sounds. In order to mock create_bucket(), we have to
have user->create_bucket() return a MockUser, not the default nullptr that the
mock would normally return. This in turn is made more complex by the fact that
create_bucket() isn't mocked normally because it takes so many parameters.
Luckily Copilot helped with that.

Next: Start work on similar mocks for Object operations.