Checkpoint: Setting up for gRPC unit tests.
Add a barebones .proto and set up all the usual unit test nonsense. Let's go
with the C++ unit test gRPC server to start with this time, instead of the
Python server - the functionality is simple enough.

The server just maintains in-memory maps of buckets and objects, each with their own
attribute maps.