Checkpoint: Working gRPC test server.
Simplify the protocol and match to the in-tree attributes a little. In
particular, we no longer have DeleteBucketAttributes and
DeleteObjectAttributes RPCs; instead, the Set*Attributes RPCs take both
attributes to add and attributes to delete.

About to split the gRPC server into a new header, the unit test source is
getting huge.

Next is to add gRPC initialisation and callouts in the actual filter.