Checkpoint: Starting gRPC client integration.
Fixed serious error in MDOffloadFilterDriver::initialize(), which wasn't
calling the superclass initialize() method. Need to check for this every time.

For every custom type we create, we need to pass the MDOffloadFilterDriver* to
its constructor so we can get back to the driver instance, which has the gRPC
infrastructure. This is somewhat akin to the HandoffHelper in Handoff.

Trying to not create a heavy gRPC client wrapper class like Handoff. I'm not sure
it adds a lot of value, and it may be overkill for our needs. Copilot pretty
much autocompletes most of the gRPC boilerplate for us anyway.