Checkpoint: Strict mock, rework with gRPC present.
Switch to StrictMock instead of NiceMock. Since our point is to be very,
*very* careful about what the driver passes through, this seems like one
of those scenarios where StrictMock applies. It will make the tests more
brittle.

As the gRPC integration is added, more and more of the tests are going
to exercise gRPC functionality. We have a server for that, so modify the
harness so it instantiates the filter driver to point at a running test
server that automatically creates buckets and objects.