Implement <driver>::get_object().
This is called by RGWHandler_REST_S3::init_from_header() to initialise an
object with no bucket reference. It still needs to return the right type
of object, but we need to be careful as there's no bucket so we can't do
any lookups on it.

The object might get filled out later as the request is processed.

Without this override, we end up with a FilterObject instead of an
MDOffloadObject, which then fails later when we try to access MDOffload-specific
functionality.