Format and constructor tweaks.
We don't want to be putting raw attribute values into the logs, as they
can be binary data. So change the logging to show the length of the
attribute values instead. Over time we can improve this to decode the
values based on the attribute name, but for now this is safer.

Add a fmtlib formatter for rgw_obj_key.