Start the pivot from GMock.
GMock is not actually that well-suited to the SAL layout. It's just too
brittle, even with NiceMock, because we're creating mock 'subobjects' all the
time. I've been finding it harder to write tests than to actually write the
code to do the work, and that's a bad bad sign.

(It's normal for tests to be a lot /more/ code than the CoT, but it's
shouldn't be harder to write!)

So: I had Copilot (with GPT-5.1-Codex (Preview)) build me an implementation of
the rgw_sal.h interface. Copilot again did a nice job, though I had to make
some minimal changes (one line).