The Great Mock Replacement.
All the mock tests are now replaced with tests using the fake driver. Now, we
have a concrete class that does very little, but can be easily modified to do
what we want for testing.

The mock SAL was becoming unmaintainable very quickly. The moment I added the
gRPC client to the filter driver, the number of expectations started to
balloon and it was not only tiresome and genuinely difficult to do, but it was
also extremely implementation-dependent and so brittle. In the end, it was an
easy decision to replace it.

Next change will remove the mock traces.