Initial create_bucket() and Bucket attr impl.
After all the test modifications, finally commit changes to the MDOffload
filter driver that implement create_bucket() (on User) and bucket attribute
storage.

This is actually working with the Python remote store already. All the test
changes were required because the mocks *exploded* the moment I started having
to call out to subobjects to fetch bucket IDs and the like.

Nonetheless, this is a significant step forward: We're able to create buckets,
and persistently set and retrieve bucket attributes from the remote store.