Object attributes get and set.
gRPC get and set working, with an important caveat: There's no explicit
object ID the way there is for buckets, so it's not clear how we
distinguish between different 'generations' of an object with the same
name. In a versioned bucket this doesn't matter, but in a nonversioned
bucket we need a way to delete an object's attributes.

The get and set looks undramatic: It works the same way as for bucket
attributes with no obvious caveats at this time.

Delete looks like we'll have to subclass FilterObject::DeleteOp as well,
which is some plumbing but not difficult. We'll probably visit
FilterObject::ReadOp at the same time as that's on the todo list for
get_attr() calls, we can at least get some logging around it.