Debug attrs set outside SAL API.
Seems we're only seeing a small subset of the attrs that actually get
set on an object. So thanks for that.

Looks like we'll have to hook rgw::sal::Writer to see all the things
that get set before the object is written. This includes important
things like the ETag and the content-type.