Checkpoint: Implement rgw::sal::Writer subclass.
Turns out that on write the xattrs for an object are written to disk
using Writer::complete(), with the writer obtained using
driver->get_append_writer() or driver->get_atomic_writer().

Subclass this (thanks Copilot) and implement with a gRPC write for the
full set of attributes. This is progress; the attrs uploaded via the
various *attrs() mutators are incomplete.