Checkpoint. FilterLog, better attrs, proto 'new'.
Use the FilterLog* instead of Filter* as base class. Get logging all the
time, so we don't have to use breakpoints which is a PITA.

Be much more careful about attr export and import. This is focused
around Writer::complete() and ReadOp::prepare() for the moment. I
believe we'll be able to extract some getter and setter code elsewhere
as these methods are the most important for most things.
(get/put-object-tagging obviously excepted.)

More work is needed on deciding which tags to export for writes and to
import for reads. It's currently working, but SSE-S3 is worth testing
now.

Add a protobuf bool for setting attributes that tells the server this is
a new object. This allows the server to delete existing tags for this
object key. I checked, S3 tags don't stay with the object key, they're
overwritten.