Checkpoint: Handle crypto object attrs.
Seems to be working (for simple ops anyway) with crypto, even when we're
filtering the attributes and preventing them being written to Rados.
This is quite a big win, I expected a fight.

This appears to mean that the crypto attrs are being read in a sane way.

Next challenge: Check the conditional write stuff still works. I suspect
it will, as I'm not externalising dates (yet).