Abstract out use of the logging filter driver.
The FilterLog* classes are super useful during development, but we
definitely don't want an l20 logging wrapper class in production code.
Make it a preprocessor switch (RGW_MDOFFLOAD_LOGGING_FILTER_DRIVER) to
use the logging or normal variants.

After the single preprocessor definition (sorry Bjarne, maybe we can do
better when we get code generation in c++, what, 26?) we use using to
define which superclass we want. We'll have to keep the list of
superclasses updated as more are added.