Versioning start. Hack for MP Etag.
Start work on adding a schema version field.

Now: There's a bit of a problem with writing MP objects. For regular
objects, we get all the tags at Writer::complete() (I think). At least
we definitely get the ETag.

For MultipartUpload::complete(), though, we don't get the etag. That's
calculated in the upstream implementation by hashing all the parts. Fair
enough, but it means that we can't push the attributes before calling
the upstream, which means we have a problem if the gRPC call fails for
some reason - we have an object in the bucket, but no attributes. That's
potentially very serious if, for example, we lose the crypto attributes.

We have to have a rollback, IOW. That's not great.