Make bucket attrs deletable. Sync bucket attrs.
rgw::sal::Bucket::merge_and_store_attrs() doesn't work the way the
documentation comment suggests. It ends up overwriting whatever attributes are
currently set with whatever's in the new_attrs argument. It's confusing, sure,
but since there's no method to delete attributes it's the only way to perform
operations like, say, deleting tags.

Add a boolean to the SetBucketAttributesRequest message to indicate that the
existing attributes should be replaced with the new set. Update
MDOffloadBucket::merge_and_store_attrs() to set that boolean when making the
gRPC call.

Remove MDOffloadBucket's cached_attrs_ member. It was a good idea and may
return when we start caching things more aggressivly, but for now use the
parent objects' attrs fields for, er, 'clarity'. Bucket doesn't seem to have
the same tricky relationship with its API that Object has, so it's less
problematic to use the parent class's attrs field. Or so it seems.

Fix some bad log messages that referred to 'variant 3' when they should have
referred to 'variant 1'. That confused the hell out of me while debugging.